+++ src/sim_search_part_patterns.hpp
@@ -49,20 +49,20 @@ inline void check_part(
       if (trim_direction == TrimDirection::Mid || (trim_direction == TrimDirection::End && metric == 'H')) {
         if (trim_direction == TrimDirection::Mid) {
           MidTrimFunc midTrim = getMidTrimFunc(metric);
-          for (int i = 0; i < string_indeces->size(); i++)
+          for (size_t i = 0; i < string_indeces->size(); i++)
             trimmed_strings[i] = midTrim(strings[string_indeces->at(i)], entry->first);
         } else
-          for (int i = 0; i < string_indeces->size(); i++)
+          for (size_t i = 0; i < string_indeces->size(); i++)
             trimmed_strings[i] = trimString<trim_direction>(strings[string_indeces->at(i)], part_len);
 
-        for (int i = 0; i < string_indeces->size(); i++) {
+        for (size_t i = 0; i < string_indeces->size(); i++) {
           std::string trim_str1 = trimmed_strings[i];
-          int str_idx1 = string_indeces->at(i);
+          size_t str_idx1 = string_indeces->at(i);
           std::string str1 = strings[str_idx1];
           local_out.insert({str_idx1, str_idx1});
-          for (int j = i + 1; j < string_indeces->size(); j++) {
+          for (size_t j = i + 1; j < string_indeces->size(); j++) {
             std::string trim_str2 = trimmed_strings[j];
-            int str_idx2 = string_indeces->at(j);
+            size_t str_idx2 = string_indeces->at(j);
             std::string str2 = strings[str_idx2];
             if (distance_k(trim_str1, trim_str2, cutoff) && distance_k(str1, str2, cutoff)) {
               if (str_idx1 > str_idx2)
@@ -73,16 +73,16 @@ inline void check_part(
           }
         }
       } else {
-        for (int i = 0; i < string_indeces->size(); i++)
+        for (size_t i = 0; i < string_indeces->size(); i++)
           trimmed_strings[i] = trimString<trim_direction>(strings[string_indeces->at(i)], part_len);
         
-        for (int i = 0; i < string_indeces->size(); i++) {
+        for (size_t i = 0; i < string_indeces->size(); i++) {
           std::string trim_str1 = trimmed_strings[i];
-          int str_idx1 = string_indeces->at(i);
+          size_t str_idx1 = string_indeces->at(i);
           local_out.insert({str_idx1, str_idx1});
-          for (int j = i + 1; j < string_indeces->size(); j++) {
+          for (size_t j = i + 1; j < string_indeces->size(); j++) {
             std::string trim_str2 = trimmed_strings[j];
-            int str_idx2 = string_indeces->at(j);
+            size_t str_idx2 = string_indeces->at(j);
             if (distance_k(trim_str1, trim_str2, cutoff)) {
               if (str_idx1 > str_idx2)
                 local_out.insert({str_idx2, str_idx1});
